requirements.txt
bcrypt==4.3.0
cloudinary==1.44.1
fastapi==0.115.13
Jinja2==3.1.6
passlib==1.7.4
pymongo==4.13.2
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
starlette==0.46.2
uvicorn==0.34.3
